+++ Addon/ElvUI_mMediaTag/modules/portraits/arena.lua
@@ -18,7 +18,7 @@ function module:InitializeArenaPortrait()
 
 				if portraits[unit] then
 					portraits[unit].__owner = parent_frame
-					portraits[unit].unit = parent_frame.unit or unit
+					portraits[unit].unit = parent_frame.__unit or unit
 					portraits[unit].type = type
 					portraits[unit].db = E.db.mMediaTag.portraits.arena
 					portraits[unit].size = E.db.mMediaTag.portraits.arena.size